requirements-dev.txt
autopep8
mypy
sqlalchemy-stubs
pylint
pytest
pytest-cov